+++ nfc/ev2/auth_test.go
@@ -1,4 +1,4 @@
-package ntag424
+package ev2
 
 import (
 	"bytes"